split helperFunctions in separate files
